refactored old schema tests to fit inside unittests. Also: added github actions.
