+++ README.rst
@@ -332,23 +332,23 @@ Test cornflow
 
 Then you have to run the following commands::
 
-    export FLASK_APP=cornflow.app
     export FLASK_ENV=testing
 
 Finally you can run all the tests with the following command::
 
-    coverage run  --source=./cornflow/ -m unittest discover -s=./cornflow/tests/
+    python -m unittest discover -s cornflow.tests
 
 If you want to only run the unit tests (without a local airflow webserver)::
 
-    coverage run  --source=./cornflow/ -m unittest discover -s=./cornflow/tests/unit/
+    python -m unittest discover -s cornflow.tests.unit
 
 If you want to only run the integration test with a local airflow webserver::
 
-    coverage run  --source=./cornflow/ -m unittest discover -s=./cornflow/tests/integration/
+    python -m unittest discover -s cornflow.tests.integration
 
 After if you want to check the coverage report you need to run::
 
+    coverage run  --source=./cornflow/ -m unittest discover -s=./cornflow/tests/
     coverage report -m
 
 or to get the html reports::
